Completed QR Scanning functionality
